+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectRelationship/EditObjectRelationship.tsx
@@ -13,7 +13,7 @@ import {
 import React from 'react';
 
 import {EditObjectRelationshipContent} from './EditObjectRelationshipContent';
-import {useObjectRelationshipForm} from './ObjectRelationshipFormBase';
+import {useObjectRelationshipForm} from './useObjectRelationshipForm';
 
 interface EditObjectRelationshipProps {
 	baseResourceURL: string;
